renaming io_sockets to io_overview, changing status_bulb to a ValueElement
