+++ field_friend/interface/components/development.py
@@ -4,7 +4,7 @@ import rosys
 from nicegui import ui
 
 from .hardware_control import hardware_control
-from .io_sockets import io_sockets
+from .io_overview import io_overview
 from .status_dev import status_dev_page
 
 if TYPE_CHECKING:
@@ -28,8 +28,5 @@ def development(system: 'System') -> None:
                             rosys.simulation_ui()
                     hardware_control(system.field_friend, system.automator, system.puncher)
                     status_dev_page(system.field_friend, system)
-    with ui.row().style('width: calc(100vw - 2rem); flex-wrap: nowrap;'):
-        with ui.card().style('background-color: #2E5396; color: white;'), ui.row():
-            with ui.row().style("width: 100%"):
-                ui.label("I/O Overview").style('font-size: 1.5rem; color: white;')
-            io_sockets(system)
+    with ui.row().style('width: calc(100vw - 2rem); flex-wrap: nowrap;'):        
+        io_overview(system)
